Adding option for icon theme
